Created different user types
Created the abstract class User, with name, username, password variables, and login() and logout() methods.

Created Administrator, Professor, and Student classes. These different types of users will extend the superclass User, but will not have overlapping permissions and abilities.

One potential change we could make is separating the administrators and professors from students by creating a Faculty class. I am unsure if we are going to need Professors and Admins to be able to have certain permissions that faculty in general should have.